+++ daac/variables.tf
@@ -17,6 +17,26 @@ variable "MATURITY" {
   default = "dev"
 }
 
+variable "bucket_config" {
+  type = map(object({
+    type = string
+    oai  = optional(string)
+  }))
+  default     = {}
+  description = "Maturity specific overrides for the base bucket config."
+}
+
+# type = {"standard", "protected", "public", "workflow"}
+# oai - The OAI ID will be added to the bucket policy to allow data distribution via CloudFront for this bucket.
+variable "bucket_config_base" {
+  type = map(object({
+    type = string
+    oai  = optional(string)
+  }))
+  default     = {}
+  description = "Map of buckets to create. Each bucket has a config that can be used to set the bucket type and enable extra features on the bucket. Add new features here as necessary."
+}
+
 variable "cma_version" {
   type        = string
   description = "Cumulus Message Adapter release version from https://github.com/nasa/cumulus-message-adapter/releases."
@@ -28,28 +48,10 @@ variable "dashboard_cloudfront_oai_id" {
   description = "CloudFront OAI ID to use for dashboard bucket policy."
 }
 
-variable "distribution_bucket_oais" {
-  type        = map(string)
-  default     = {}
-  description = "A map of bucket names to CloudFront OAI IDs. The OAI IDs will be added to the bucket policy to allow data distribution via CloudFront for those buckets."
-}
-
 variable "partner_bucket_names" {
   type        = list(string)
   default     = []
   description = "List of buckets which we need access to but do not create. Include the full bucket name."
-}
-
-variable "protected_bucket_names" {
-  type        = list(string)
-  default     = []
-  description = "List of 'protected' buckets to create. The stack prefix is automatically added to the bucket names."
-}
-
-variable "public_bucket_names" {
-  type        = list(string)
-  default     = []
-  description = "List of 'public' buckets to create. The stack prefix is automatically added to the bucket names."
 }
 
 variable "s3_replicator_target_bucket" {
@@ -62,16 +64,4 @@ variable "s3_replicator_target_prefix" {
   type        = string
   default     = null
   description = "Prefix that the S3 replicator will write logs to in the target bucket."
-}
-
-variable "standard_bucket_names" {
-  type        = list(string)
-  default     = []
-  description = "List of 'standard' buckets to create. The stack prefix is automatically added to the bucket names."
-}
-
-variable "workflow_bucket_names" {
-  type        = list(string)
-  default     = []
-  description = "List of 'workflow' buckets to create. The stack prefix is automatically added to the bucket names."
 }